refactor cardsPage add headingOfPage, refactor BackBtn with e.preventDefault
