+++ src/Pages/CardsPage/Cards.page.tsx
@@ -1,13 +1,8 @@
-import { ChangeEvent, useContext, useState } from 'react'
+import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { Link, useParams } from 'react-router-dom'
+import { useParams } from 'react-router-dom'
 
-import groupIcon from '@/assets/icons/WhiteSVG/Group 1399.svg'
-import groupIconBlack from '@/assets/icons/WhiteSVG/Group 1399Black.svg'
-import menuIcon2 from '@/assets/icons/WhiteSVG/edit-2-outline.svg'
-import playIcon from '@/assets/icons/WhiteSVG/play-circle-outline.svg'
-import menuIcon from '@/assets/icons/WhiteSVG/trash-outline.svg'
-import { handleToastInfo } from '@/common/consts/toastVariants'
+import { HeadingOfPage } from '@/Pages/CardsPage/HeadingSecondRow/HeadingOfPage'
 import { headersNameCards, initCurrentPage, selectOptionPagination } from '@/common/globalVariables'
 import { ModalAddEditDeck } from '@/components/Modals/ModalAddEditDeck/ModalAddEditDeck'
 import { DeleteModal } from '@/components/Modals/ModalDelete/DeleteModal'
@@ -15,17 +10,12 @@ import { ModalAddEditCard } from '@/components/Modals/ModalEditCard/ModalAddEdit
 import ModalOnEmpty from '@/components/Modals/ModalOnEmpty/ModalOnEmpty'
 import { SingleRowCard } from '@/components/TableComponent/SingleRowCard/SingleRowCard'
 import { TableComponentWithTypes } from '@/components/TableComponent/TableComponentWithTypes'
-import { BackBtn } from '@/components/ui/BackBtn/BackBtn'
-import DropdownMenuDemo from '@/components/ui/DropDown/DropDown'
-import DropDownItem from '@/components/ui/DropDown/DropDownItem'
-import Input from '@/components/ui/Input/Input'
 import Loading from '@/components/ui/Loading/Loading'
 import { LoadingBar } from '@/components/ui/LoadingBar/LoadingBar'
 import { Page } from '@/components/ui/Page/Page'
 import { PaginationWithSelect } from '@/components/ui/Pagination/PaginationWithSelect'
 import Typography from '@/components/ui/Typography/Typography'
 import { Button } from '@/components/ui/button'
-import { UserContext } from '@/components/ui/changeTheme/Context'
 import { useQueryParams } from '@/hooks/useQueryParams'
 import { path } from '@/router/path'
 import { router } from '@/router/router'
@@ -33,22 +23,18 @@ import { useMeQuery } from '@/services/auth/auth.service'
 import { useDeleteCardByIdMutation, useGetCardsQuery } from '@/services/cards/cards.service'
 import { CardResponse } from '@/services/cards/cards.types'
 import { useDeleteDeckMutation, useGetDeckByIdQuery } from '@/services/decks/decks.service'
-import { clsx } from 'clsx'
 
 import s from './cardsPage.module.scss'
 
 export const CardsPage = () => {
-  const context = useContext(UserContext)
   const { t } = useTranslation()
   const {
     currentOrderBy,
     currentPage,
     debouncedSearchValue,
     itemsPerPage,
-    search,
     setCurrentPageQuery,
     setItemsPerPageQuery,
-    setSearchQuery,
   } = useQueryParams()
 
   const [deleteCard] = useDeleteCardByIdMutation()
@@ -82,11 +68,6 @@ export const CardsPage = () => {
     setCurrentPageQuery(value)
   }
 
-  const handleSearch = (e: ChangeEvent<HTMLInputElement>) => {
-    setCurrentPageQuery(Number(initCurrentPage))
-    setSearchQuery(e.currentTarget.value)
-  }
-
   const onDeleteDeckHandler = () => {
     deleteDeck({ id: deckId })
     setIsDeleteDeckModal(true)
@@ -100,24 +81,9 @@ export const CardsPage = () => {
 
   const cardsData = currentData ?? data
 
-  const isCardsCountFilled = currentData?.items.length !== 0
-
   const isCardsCountZero = currentData?.items.length === 0
 
   const isMineCards = currentDeckData?.userId === meData?.id
-
-  const handleOpenModal = () => {
-    if (isMineCards && isCardsCountZero) {
-      setIsEmptyModal(true)
-    } else {
-      console.log('baga')
-      router.navigate(`${path.decks}`)
-    }
-  }
-
-  const notifyLearnHandler = () => {
-    handleToastInfo(`Add card before learning!`)
-  }
 
   const loadingStatus = isFetching || isDeckLoading
 
@@ -155,85 +121,16 @@ export const CardsPage = () => {
           <Typography variant={'h1'}>{cardItem?.question}</Typography>
           <Typography variant={'body1'}>{t('cardsPage.isDeleteCard')}</Typography>
         </DeleteModal>
-        <div className={s.heading}>
-          <BackBtn onClick={handleOpenModal} to={'#'}>
-            {t('cardsPage.backDeckList')}
-          </BackBtn>
-          <div className={s.headingSecondRow}>
-            <div className={clsx(deck?.cover && s.isWithImage)}>
-              <div className={s.info}>
-                <Typography as={'h1'} variant={'h1'}>
-                  {deck?.name}
-                </Typography>
-                {isMineCards && (
-                  <DropdownMenuDemo
-                    className={s.dropdown}
-                    icon={context?.theme === 'moon' ? groupIcon : groupIconBlack}
-                    type={'menu'}
-                  >
-                    {isCardsCountZero ? (
-                      <DropDownItem
-                        handleOnClick={notifyLearnHandler}
-                        icon={playIcon}
-                        text={t('cardsPage.learn')}
-                      />
-                    ) : (
-                      <DropDownItem
-                        href={`${path.decks}/${deckId}${path.learn}`}
-                        icon={playIcon}
-                        text={t('cardsPage.learn')}
-                      />
-                    )}
-
-                    <DropDownItem
-                      handleOnClick={() => setIsUpdateDeckModal(true)}
-                      icon={menuIcon2}
-                      text={t('cardsPage.edit')}
-                    />
-                    <DropDownItem
-                      handleOnClick={() => setIsDeleteDeckModal(true)}
-                      icon={menuIcon}
-                      text={t('cardsPage.delete')}
-                    />
-                  </DropdownMenuDemo>
-                )}
-              </div>
-              {deck?.cover && <img alt={'img'} src={deck?.cover} width={'120px'} />}
-            </div>
-            {isCardsCountFilled && (
-              <div className={s.switchButton}>
-                {isMineCards ? (
-                  <Button
-                    className={s.addCard}
-                    onClick={() => setIsCreateCardModal(true)}
-                    type={'button'}
-                  >
-                    <Typography variant={'subtitle2'}>{t('cardsPage.addNewCard')}</Typography>
-                  </Button>
-                ) : (
-                  <Button
-                    as={Link}
-                    className={s.learnCards}
-                    onClick={() => setIsUpdateCardModal(true)}
-                    to={`${path.decks}/${deckId}${path.learn}`}
-                    type={'button'}
-                  >
-                    <Typography variant={'subtitle2'}>{t('cardsPage.learnCards')}</Typography>
-                  </Button>
-                )}
-              </div>
-            )}
-          </div>
-          {isCardsCountFilled && (
-            <Input
-              callback={setSearchQuery}
-              className={s.input}
-              currentValue={search}
-              onChange={handleSearch}
-              type={'search'}
-            />
-          )}
-        </div>
+        <HeadingOfPage
+          deck={deck}
+          deckId={deckId}
+          isCardsCountZero={isCardsCountZero}
+          isMineCards={isMineCards}
+          openCreateCardModalHandler={setIsCreateCardModal}
+          openDeleteDeckModalHandler={setIsDeleteDeckModal}
+          openEditDeckModalHandler={setIsUpdateDeckModal}
+          openEmptyDeckModalHandler={setIsEmptyModal}
+        />
         {isCardsCountZero ? (
           <div className={s.emptyContent}>
             <Typography variant={'body1'}>
