+++ src/Pages/CardsPage/cardsPage.module.scss
@@ -1,21 +1,3 @@
-.heading {
-  display: grid;
-  grid-template-rows: auto 1fr;
-  width: 100%;
-  margin-bottom: 24px;
-}
-
-.headingSecondRow {
-  display: flex;
-  align-items: start;
-  justify-content: space-between;
-  margin-bottom: 24px;
-
-  & .isWithImage > h1 {
-    margin-bottom: 15px;
-  }
-}
-
 .emptyContent {
   display: flex;
   flex-direction: column;
@@ -36,38 +18,11 @@
   height: 36px;
 }
 
-.input {
-  width: 100%;
-}
-
 .footer {
   width: 100%;
   margin: 24px 0;
 }
 
-.switchButton {
-  display: flex;
-  align-items: end;
-  justify-content: end;
-
-  & > button:first-child > p {
-    margin: 0;
-  }
-}
-
 .learnCards {
   text-decoration: none;
-}
-
-.info {
-  display: flex;
-  gap: 6px;
-  align-items: center;
-}
-
-.dropdown {
-  display: flex;
-  align-items: center;
-  width: 24px !important;
-  height: 24px !important;
 }
